refactor: extract service supervisor
